only fall back to in-place writes when no temp file can be created
The in-place fallback now triggers only when the temporary file itself
can't be created; a failure after that point (fchmod, write, flush) is
reported rather than truncating the target first. Callers can turn the
fallback off entirely (Replace All does), and an ownerOnly in-place
write that can't restrict the file's mode now fails instead of writing
private contents into a file it couldn't make private.

Also read integer-typed pref values via getDouble() so that wide
integers don't trip getInt()'s assertion, and add a Windows test that
replaces a file held open by one of our readers.

+++ src/cpp/session/modules/SessionUserPrefs.cpp
@@ -161,7 +161,8 @@ json::Value coerceNumber(const json::Value& value, json::Type expectedType)
    }
    else if (expectedType == json::Type::REAL && value.getType() == json::Type::INTEGER)
    {
-      return json::Value(static_cast<double>(value.getInt()));
+      // getDouble() accepts every integer width; getInt() would not
+      return json::Value(value.getDouble());
    }
 
    return value;
